requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
coverage==5.3
diff-pdf-visually==1.5.1
dj-database-url==0.5.0
Django==3.1
django-heroku==0.3.1
django-nose==1.4.7
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
nose==1.3.7
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyPDF2==1.26.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
tzlocal==2.1